requirements.txt
absl-py==0.11.0
aiohttp==3.7.3
async-timeout==3.0.1
attrs==20.3.0
cachetools==4.2.1
certifi==2020.12.5
chardet==3.0.4
click==7.1.2
configparser==5.0.1
cycler==0.10.0
docker-pycreds==0.4.0
fsspec==0.8.7
future==0.18.2
gitdb==4.0.5
GitPython==3.1.13
google-auth==1.27.0
google-auth-oauthlib==0.4.2
grpcio==1.35.0
idna==2.10
joblib==1.0.1
kiwisolver==1.3.1
Markdown==3.3.4
matplotlib==3.3.4
multidict==5.1.0
numpy==1.20.1
oauthlib==3.1.0
pathtools==0.1.2
pillow==8.1.0
promise==2.3
protobuf==3.15.2
psutil==5.8.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyglet==1.5.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytorch-lightning==1.2.1
pytorch-lightning-bolts==0.3.0
PyYAML==5.3.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-learn==0.24.1
scipy==1.6.1
sentry-sdk==0.20.3
shortuuid==1.0.1
six==1.15.0
smmap==3.0.5
subprocess32==3.5.4
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
threadpoolctl==2.1.0
torch==1.7.1
torchvision==0.8.2
tqdm==4.57.0
typing-extensions==3.7.4.3
urllib3==1.26.3
wandb==0.10.20
Werkzeug==1.0.1
yarl==1.6.3
